refactor: 병합 해결
