+++ src/main/java/com/sparklenote/user/handler/CustomSuccessHandler.java
@@ -2,7 +2,7 @@ package com.sparklenote.user.handler;
 
 
 import com.sparklenote.user.jwt.JWTUtil;
-import com.sparklenote.user.oAuth.CustomOAuth2User;
+import com.sparklenote.user.oAuth2.CustomOAuth2User;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -26,7 +26,7 @@ public class CustomSuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
 
     @Value("${jwt.expiration}")
     private Long jwtExpiration;
-  
+
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
 
@@ -39,6 +39,7 @@ public class CustomSuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
         Iterator<? extends GrantedAuthority> iterator = authorities.iterator();
         GrantedAuthority auth = iterator.next();
         String role = auth.getAuthority();
+
         String token = jwtUtil.createJwt(username, role, jwtExpiration);
 
         response.addCookie(createCookie("Authorization", token));
