+++ src/main/java/com/sparklenote/user/jwt/JWTUtil.java
@@ -16,7 +16,7 @@ import java.util.Date;
 public class JWTUtil {
 
     private SecretKey secretKey;
-  
+
     public JWTUtil(@Value("${spring.jwt.secret}") String secret) {
         secretKey = new SecretKeySpec(secret.getBytes(StandardCharsets.UTF_8), SignatureAlgorithm.HS256.getJcaName());
     }
